fileware/requirements.txt
gmail-connector>=0.6.0
pyngrok>=5.2.1
pyOpenSSL>=22.0.0
python-dotenv>=0.19.2
PyYAML>=6.0
requests>=2.27.1
